[김이서] refactor: 마이페이지 코드 1차 정리
