+++ src/App.jsx
@@ -3,9 +3,9 @@ import "./App.css";
 import LandingPage from "./pages/LandingPage";
 import NotFound from "./pages/NotFound";
 import ListPage from "./pages/ListPage";
-import MyPage2 from "./pages/Mypage2";
 import ExamplePage from "./pages/ExamplePage";
-import MypageTest from "./pages/MypageTest";
+import MyPage2 from "./pages/MyPage2";
+import MyPage from "./pages/Mypage";
 
 function App() {
   return (
@@ -13,11 +13,11 @@ function App() {
       <Routes>
         <Route index path="/" element={<LandingPage />} />
         <Route path="/list" element={<ListPage />} />
-        <Route path="/mypage" element={<MyPage2 />} />
+        <Route path="/mypage" element={<MyPage />} />
         <Route path="*" element={<NotFound />} />
         {/* 예시를 위한 페이지. 추후 삭제. */}
         <Route path="/example" element={<ExamplePage />} />
-        <Route path="/test" element={<MypageTest />} />
+        <Route path="/mypagetest" element={<MyPage2 />} />
       </Routes>
     </>
   );
