+++ src/components/layouts/FavoriteIdols.jsx
@@ -1,14 +1,18 @@
 import idolDeleteBtn from "../../assets/icons/btn_delete.svg";
 import IdolProfile from "../../components/common/IdolProfile";
 
-const FavoriteIdols = ({ myFavorIdols, selectedIdols, onRemoveFavorite }) => {
+const FavoriteIdols = ({
+  myFavorIdols,
+  selectedIdolsIds,
+  onRemoveFavorite,
+}) => {
   return (
     <div>
       {myFavorIdols.length === 0 ? (
         <p>관심 아이돌이 없습니다.</p>
       ) : (
         myFavorIdols.map((idol) => {
-          const isSelected = selectedIdols.includes(idol.id);
+          const isSelected = selectedIdolsIds.includes(idol.id);
           return (
             <div key={idol.id}>
               <IdolProfile idol={idol} isSelected={isSelected} />
