+++ src/components/layouts/MyPageDesk.jsx
@@ -4,7 +4,7 @@ import ProfileChunk from "../ProfileChunk";
 import prevButton from "../../assets/images/prev_btn.png";
 import nextButton from "../../assets/images/next_btn.png";
 
-const MyPageDesk = ({ selectedIdols, onToggle }) => {
+const MyPageDesk = ({ selectedIdolsIds, onToggle }) => {
   const [list, setList] = useState([]);
   const [nextCursor, setNextCursor] = useState(null);
   const [currentCursor, setCurrentCursor] = useState(null);
@@ -67,8 +67,8 @@ const MyPageDesk = ({ selectedIdols, onToggle }) => {
             key={idol.id}
             className="ProfileChunk"
             idol={idol}
-            isSelected={selectedIdols.includes(idol.id)}
-            onClick={() => onToggle(idol.id)}
+            isSelected={selectedIdolsIds.includes(idol.id)}
+            onClick={() => onToggle(idol)}
           />
         ))}
       </div>
